% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dylanie.R
\name{dylanie_model_mchains}
\alias{dylanie_model_mchains}
\title{Paralleled full Bayesian estimation for multivariate dyads data analysis (DyLAnIE).}
\usage{
dylanie_model_mchains(formula, data, mcmc_len, chains = 1, verbose = F)
}
\arguments{
\item{formula}{an object of class "formula" similar to linear regression.
It is a symbolic description of the covariates set used in the model.}

\item{data}{a list for data input. It contains binary responses, covaraites
matrix, a constraint matrix (1 for free parameters, 0 for fixed zero) for
coefficient parameter $\phi\xi$ and pre estimated measurement parameters. See
the following example for more details.}

\item{mcmc_len}{an integer specifies the length of MCMC draws}

\item{chains}{an integer specifies the number of MCMC chains (draw at the same time)}

\item{verbose}{boolen variable, whether enable verbose mode}
}
\value{
The function returns a list with the following components:
\describe{
\item{g_draws}{A matrix containing MCMC draws for coefficient parameters $\phi_\xi$.}
\item{b_tp_draws}{A matrix containing MCMC draws for coefficient parameters $\phi_\eta1$.}
\item{b_fp_draws}{A matrix containing MCMC draws for coefficient parameters $\phi_\eta2$.}
\item{sig2_tp_draws}{A matrix containing MCMC draws for variance of $\eta_1$.}
\item{sig2_fp_draws}{A matrix containing MCMC draws for variance of $\eta_2$.}
\item{rho_draws}{A matrix containing MCMC draws for correlation of $\eta_1$ and $\eta_2$.}
\item{e12_last}{A matrix containing MCMC draws for continuous random variables $\eta_1$ and $\eta_2$.}
\item{xi_last}{A matrix containing MCMC draws for discrete random variables $\xi$.}
}
}
\description{
The program perform estimation for a general latent variable modelling framework.
This framework is intended for for the analysis of
multivariate binary data on dyads, which can also handle zero inflation and
non-equivalence of measurement.
}
\examples{
\dontrun{
dylanie_mchains_res <- dylanie_model_mchains(~female, data=aidrp_data, mcmc_len = 10, chains = 2, verbose = F)
str(dylanie_mchains_res)
}

}
\references{
Kuha, J., Zhang, S., & Steele, F. (2021). manuscript
\emph{Journal of the American Statistical Association} <doi: >.
}
