% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dylanie.R
\name{dylanie_model_simple}
\alias{dylanie_model_simple}
\title{Paralleled full Bayesian estimation for multivariate dyads data analysis (DyLAnIE).}
\usage{
dylanie_model_simple(formula, data, mcmc_len, verbose = F)
}
\arguments{
\item{formula}{an object of class "formula" similar to linear regression.
It is a symbolic description of the covariates set used in the model.}

\item{data}{a list for data input. It contains binary responses, covaraites
matrix, a constraint matrix (1 for free parameters, 0 for fixed zero) for
coefficient parameter $\phi\xi$ and pre estimated measurement parameters. See
the following example for more details.}

\item{mcmc_len}{an integer specifies the length of MCMC draws}

\item{verbose}{boolen variable, whether enable verbose mode}
}
\value{
The function returns a list with the following components:
\describe{
\item{g_draws}{A matrix containing MCMC draws for coefficient parameters $\phi_\xi$.}
\item{b_tp_draws}{A matrix containing MCMC draws for coefficient parameters $\phi_\eta1$.}
\item{b_fp_draws}{A matrix containing MCMC draws for coefficient parameters $\phi_\eta2$.}
\item{sig2_tp_draws}{A matrix containing MCMC draws for variance of $\eta_1$.}
\item{sig2_fp_draws}{A matrix containing MCMC draws for variance of $\eta_2$.}
\item{rho_draws}{A matrix containing MCMC draws for correlation of $\eta_1$ and $\eta_2$.}
\item{e12_last}{A matrix containing MCMC draws for continuous random variables $\eta_1$ and $\eta_2$.}
\item{xi_last}{A matrix containing MCMC draws for discrete random variables $\xi$.}
}
}
\description{
The program perform estimation for a general latent variable modelling framework.
This framework is intended for for the analysis of
multivariate binary data on dyads, which can also handle zero inflation and
non-equivalence of measurement.
}
\examples{
\dontrun{
# load the simulated dataset
load("jags.rda")
aidrp_data_simple <- list(ytp=jags.data$ytp, yfp = jags.data$yfp, xi=xi.start,
                          female = jags.data$female,
                          distlong = jags.data$distlong,
                          age40 = jags.data$age40,
                          partner = jags.data$partner,
                          clt2 = jags.data$clt2,
                          c2to4 = jags.data$c2to4,
                          c5to10 = jags.data$c5to10,
                          c11to16 = jags.data$c11to16,
                          cgt16 = jags.data$cgt16,
                          jbs3 = jags.data$jbs == 3,
                          jbs4 = jags.data$jbs == 4,
                          page70 = jags.data$page70,
                          plive = jags.data$plive \%in\% c(3, 4, 5, 7, 9))

dylanie_res <- dylanie_model_simple(~female+age40, data=aidrp_data_simple, mcmc_len = 10, verbose = F)
}

}
\references{
Kuha, J., Zhang, S., & Steele, F. (2021). manuscript
\emph{Journal of the American Statistical Association} <doi: >.
}
