% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dylanie.R
\name{dylanie_model_update}
\alias{dylanie_model_update}
\title{Paralleled full Bayesian estimation for multivariate dyads data analysis (DyLAnIE).}
\usage{
dylanie_model_update(dylanie_res, mcmc_len, verbose = F)
}
\arguments{
\item{dylanie_res}{object return from \code{dylanie_model()} or \code{dylanie_model_simple()} function}

\item{mcmc_len}{an integer specifies the length of MCMC draws}

\item{verbose}{boolen variable, whether enable verbose mode}
}
\value{
The function returns a list with the following components:
\describe{
\item{g_draws}{A matrix containing MCMC draws for coefficient parameters $\phi_\xi$.}
\item{b_tp_draws}{A matrix containing MCMC draws for coefficient parameters $\phi_\eta1$.}
\item{b_fp_draws}{A matrix containing MCMC draws for coefficient parameters $\phi_\eta2$.}
\item{sig2_tp_draws}{A matrix containing MCMC draws for variance of $\eta_1$.}
\item{sig2_fp_draws}{A matrix containing MCMC draws for variance of $\eta_2$.}
\item{rho_draws}{A matrix containing MCMC draws for correlation of $\eta_1$ and $\eta_2$.}
\item{e12_last}{A matrix containing MCMC draws for continuous random variables $\eta_1$ and $\eta_2$.}
\item{xi_last}{A matrix containing MCMC draws for discrete random variables $\xi$.}
}
}
\description{
The program perform estimation for a general latent variable modelling framework.
This framework is intended for for the analysis of
multivariate binary data on dyads, which can also handle zero inflation and
non-equivalence of measurement.
}
\examples{
\dontrun{
dylanie_res_updated <- dylanie_model_update(dylanie_res, 10, verbose = F)
}
}
\references{
Kuha, J., Zhang, S., & Steele, F. (2021). manuscript
\emph{Journal of the American Statistical Association} <doi: >.
}
